Added screenshot and minor change
Minor change was changing a uint to uint256 in the KaseiCoinCrowdale contract

+++ KaseiCoinCrowdsale.sol
@@ -12,7 +12,7 @@ contract KaseiCoinCrowdsale is Crowdsale, MintedCrowdsale {
     
     // Provide parameters for all of the features of the crowdsale, `rate`, `wallet` for fundraising, and `token`.
     constructor(
-        uint rate,
+        uint256 rate,
         address payable wallet,
         KaseiCoin token
     ) public Crowdsale(rate, wallet, token) {
